pyproject.toml
[tool.poetry]
name = "dealflow"
version = "0.1.0"
description = "A simple Freelance API using Flask and its front-end using React"
authors = ["Ahmed Emad <50369848+ahmedemad242@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dealflow", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8.1"
flask = "^2.3.2"
flask-cors = "^3.0.10"
flask-migrate = "^4.0.4"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.3"
sqlalchemy = "^2.0.16"
psycopg2-binary = "^2.9.6"
flask-bcrypt = "^1.0.1"
cryptography = "^41.0.1"
apifairy = "^1.3.0"
flask-marshmallow = "^0.15.0"
click = "^8.1.3"
faker = "^18.11.0"
marshmallow-sqlalchemy = "^0.29.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.3.3"
pytest = "^7.3.2"
pytest-clarity = "^1.0.1"
pytest-dotenv = "^0.5.2"
pytest-flask = "^1.2.0"
tox = "^4.6.2"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 89
target-version = ['py37']
include = '\.pyi?$'
exclude =  '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.vscode
    | __pycache__
    | _build
    | buck-out
    | build
    | dist
    | venv
)/
'''
